# AspireTemplate.ApiService/Domain/models/Error.cs
// Copyright (c) Microsoft Corporation. All rights reserved.
// Licensed under the MIT License.
// <auto-generated />

using System;
using System.Text.Json;
using System.Text.Json.Serialization;

namespace PersonAPI.Service.Models
{

    public partial class Error
    {
        public string RequestId { get; set; }

        public string TraceId { get; set; }

        public string Detail { get; set; }

        public string Title { get; set; }

        public string Type { get; set; }


    }
}