# AspireTemplate.ApiService/Domain/models/ErrorDetail.cs
// Copyright (c) Microsoft Corporation. All rights reserved.
// Licensed under the MIT License.
// <auto-generated />

using System;
using System.Text.Json;
using System.Text.Json.Serialization;

namespace PersonAPI.Service.Models
{

    public partial class ErrorDetail
    {
        public string Key { get; set; }

        public string[] Messages { get; set; }


    }
}