# AspireTemplate.ApiService/Domain/models/InternalServerError.cs
// Copyright (c) Microsoft Corporation. All rights reserved.
// Licensed under the MIT License.
// <auto-generated />

using System;
using System.Text.Json;
using System.Text.Json.Serialization;

namespace PersonAPI.Service.Models
{

    public partial class InternalServerError : Error
    {
        public object StatusCode { get; set; }


    }
}