# AspireTemplate.ApiService/openapi3.yaml
openapi: 3.0.0
info:
  title: Person API
  description: This is a pet store service.
  version: 0.0.0
tags: []
paths:
  /persons:
    post:
      operationId: PersonOperations_createPerson
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Person'
        '400':
          description: The server could not understand the request due to invalid syntax.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '401':
          description: Access is unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
        '403':
          description: Access is forbidden.
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/ForbiddenError'
                  - $ref: '#/components/schemas/ForbiddenError'
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundError'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                person:
                  $ref: '#/components/schemas/Person'
              required:
                - person
  /persons/:
    get:
      operationId: PersonOperations_listPersons
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Person'
        '400':
          description: The server could not understand the request due to invalid syntax.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '401':
          description: Access is unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
        '403':
          description: Access is forbidden.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenError'
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundError'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
  /persons/{id}:
    get:
      operationId: PersonOperations_getPerson
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Person'
        '400':
          description: The server could not understand the request due to invalid syntax.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '401':
          description: Access is unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
        '403':
          description: Access is forbidden.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenError'
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundError'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
    put:
      operationId: PersonOperations_updatePerson
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Person'
        '400':
          description: The server could not understand the request due to invalid syntax.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '401':
          description: Access is unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
        '403':
          description: Access is forbidden.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenError'
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundError'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                person:
                  $ref: '#/components/schemas/Person'
              required:
                - person
    delete:
      operationId: PersonOperations_deletePerson
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: The request has succeeded.
        '400':
          description: The server could not understand the request due to invalid syntax.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '401':
          description: Access is unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
        '403':
          description: Access is forbidden.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenError'
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundError'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
components:
  schemas:
    Error:
      type: object
      required:
        - requestId
        - traceId
        - detail
        - title
        - type
      properties:
        requestId:
          type: string
        traceId:
          type: string
        detail:
          type: string
        title:
          type: string
        type:
          type: string
    ErrorDetail:
      type: object
      required:
        - key
        - messages
      properties:
        key:
          type: string
        messages:
          type: array
          items:
            type: string
    ForbiddenError:
      type: object
      required:
        - statusCode
      properties:
        statusCode:
          type: number
          enum:
            - 403
      allOf:
        - $ref: '#/components/schemas/Error'
    InternalServerError:
      type: object
      required:
        - statusCode
      properties:
        statusCode:
          type: number
          enum:
            - 500
      allOf:
        - $ref: '#/components/schemas/Error'
    NotFoundError:
      type: object
      required:
        - statusCode
      properties:
        statusCode:
          type: number
          enum:
            - 404
      allOf:
        - $ref: '#/components/schemas/Error'
    Person:
      type: object
      required:
        - id
        - name
        - age
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        age:
          type: integer
          format: int32
          minimum: 0
          maximum: 150
    UnauthorizedError:
      type: object
      required:
        - statusCode
      properties:
        statusCode:
          type: number
          enum:
            - 401
      allOf:
        - $ref: '#/components/schemas/Error'
    ValidationError:
      type: object
      required:
        - statusCode
        - errors
      properties:
        statusCode:
          type: number
          enum:
            - 400
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ErrorDetail'
      allOf:
        - $ref: '#/components/schemas/Error'
servers:
  - url: https://localhost:7425
    description: Local development server
    variables: {}